requirements.txt
fastjsonschema >= 2.14.4
google-cloud-bigquery >= 1.24.0